requirements/common.txt
numpy>=1.23.2,<1.24
pandas>=1.4.4,<1.5
matplotlib>=3.5.3,<3.6
seaborn>=0.11.2,<0.12
python-dotenv>=0.21.0,<0.22
plotly>=5.10.0,<5.11
scikit-learn>=1.1.2,<1.2
pyarrow>=9.0.0,<9.1
streamlit>=1.13.0,<1.14
joblib>=1.2.0,<1.3
Pillow>=9.2.0,<9.3
ccxt>=1.93.108,<1.94
pandas-ta>=0.3.14b,<0.4
